requirements.txt
streamlit==1.38.0
openai==1.40.0
pydantic==2.9.2
pypdf==4.3.1
python-docx==1.1.2
python-dotenv
httpx==0.27.2
anthropic
pyautogen
